pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "paper2agent"
version = "0.1.0"
description = "Paper2AgentSkills with R-ASAA: Robust Adaptive Skill-Accretion Architecture"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.3.0",
    "chromadb>=0.4.0",
    "docling>=2.4.0",
    "pypdf>=3.0.0",
    "fpdf",
    "pydantic>=2.0",
    "gradio>=6.1.0",
    "huggingface-hub>=0.36.0",
    "mlx-llm>=1.0.9",
    "mlx-lm>=0.29.1",
]

[project.scripts]
paper2agent = "paper2agent.cli:main"

[tool.setuptools]
packages = ["paper2agent"]
